kernel_CrosECSysfs should only run against Chrome OS ECs.
This test checks the ability for userspace to talk to the embedded
controller via sysfs. That's only a feature for systems that have a Chrome
OS EC. We want to test on all of those systems, but we don't want to
enumerate them manually (because we'll be sure to miss one as new systems
are developed).

This CL adds a dependency to see if the system has a Chrome OS EC
attached. If it doesn't, the test shouldn't be run.

These are today's shipping platforms and the results of running
"mosys ec info" on each one:

  alex          smsc | mec1308 | 01VFA617
  butterfly     ENE | KB3940 | 820DG1
  daisy         stm | stm32f100 | snow_v1.3.139-375eb9f
  falco         ti | Unknown-10de | falco_v1.5.84-bed5e3e
  leon          ti | Unknown-10de | leon_v1.5.140-75d6bd8
  link          ti | lm4fs1gh5bb | link_v1.2.120-1137a99
  lumpy         smsc | mec1310 | 02WQA015
  mario         ite | it8500 | 0222
  panther       <Exit code 22>
  parrot        ENE | KB932 | 00BE107A00
  parrot_ivb    ENE | KB932 | 00BE107A00
  peppy         ti | Unknown-10de | peppy_v1.5.114-5d52788
  stout         ite | it8518 | 3.08
  stumpy        <Exit code 22>
  wolf          ti | Unknown-10de | wolf_v1.5.110-75e8e38
  ZGB           nuvoton | npce781 | 1.00

The systems with our EC use our naming scheme for the version. We'll use
that to identify them.

BUG=chromium:334014
TEST=manual (for now)

I've tested the EC-detection code (with it in kernel_CrosECSysfs.py) on all
those types, plus monroe, peach_pi, peach_pit, and rambi using test_that. It
worked correctly.

I then moved the EC-detection code into cros_host.py and ran it on mario,
stumpy, lumpy, and link using this:

  #!/usr/bin/env python
  import common
  from autotest_lib.server.hosts import factory

  host = factory.create_host('chromeos2-row3-rack6-host1.cros')
  labels = host.get_labels()

  print '\n\n\nHEY\n'
  print labels

That worked, too.

Change-Id: I5dc90bcfa5c5866233f8375b89fb6942099beaa1
Signed-off-by: Bill Richardson <wfrichar@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/186650
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>