Modified belkinF9K and WD600 router configs.
Western digital seems to randomly toggle the radios, which disable the related
fields. So now we have a check to make sure both the radios are on before we
start the test and we also wait for page to reload before we access any
dropdown field.
Multiple belkin routers use BelkinF9KAPConfigurator class for setup. Sometimes
the setup failed because the admin page wasn't loading. So now we refresh the
page if its stuck or throw an exception if the router doesnot load a page.

BUG=chromium:330274,342624,343739,343746
TEST=unittest

Change-Id: I8fed2f43f6b0e3b39791663f1ab6e5adc0e48e13
Reviewed-on: https://chromium-review.googlesource.com/186866
Reviewed-by: Deepak Gopal <deepakg@chromium.org>
Commit-Queue: Deepak Gopal <deepakg@chromium.org>
Tested-by: Deepak Gopal <deepakg@chromium.org>