Touch FW test: robot fat finger w/ resting finger
This adds the support for the Fat finger + resting finger interaction
tests in the robot wrapper for the touch FW test.  In this test the
robot uses a regular-size finger in a stationary position and then moves
the fat finger around.  This CL adds the plumbing to allow this to
happen.

BUG=chromium:334460
TEST=I ran the tests and observed the robot

Change-Id: Ie7ac488319819261033e8971c20cc8bc507c9ca7
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/187723
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>