[autotest] Include priority in AcquireHostRequest.
Before this change we would allocate hosts to competing requests,
to the request group with most demand, i.e if we had 10 requests with
board:lumpy and 5 with board:lumpy, bluetooth, we would favor
satisfying the 10 request group even if the 5 request group had
a higher priority job in it. With this change requests are grouped by
to priority, and then sorted, so higher priority jobs get to ask for
hosts first.

Downsides (that are obviously less important than correctness):
1. Same deps/acls requests will sometimes need > 1 request for hosts,
   though the priority is not a part of the actual (django) request.
2. Before, we would lock a maximum number of devices upfront with the
   highest demand request. This change leads to a slight fragmentation
   of this model.

BUG=chromium:345308
TEST=Ran 2 suites with different priorities and forced inversion without
     the change, then made sure they scheduler correctly with it.
DEPLOY=scheduler

Change-Id: Icb12cebd0c874529ccc0d199d8fe377a953da7f1
Reviewed-on: https://chromium-review.googlesource.com/187317
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>