Bluetooth SDP autotest: Service Search Request with continuation state
Added new SDP test for Service Search Request, which is an implementation of
'TP/SERVER/SS/BV-03-C' from SDP test specification. It verifies the correct
behaviour of continuation state functionality.

Continuation state is used when the whole response data to the request is too
big to fit into single response message. When this happens, the SDP server
generates a partial response along with a continuation state parameter.
The SDP client should supply this parameter to retrieve the next portion of the
complete response, and so on, until the continuation state parameter becomes
equal to zero.

Client-side support of continuation state was already implemented before in
BluetoothSDPSocket class. In order to force the DUT to actually use the
continuation state, the lower tester needs to send a request, which refers to
many services. As by default the DUT doesn't have so many services running,
the test has to create some fake services, and then request for their Class ID.
So this CL also adds a method to create Bluetooth services on the DUT, which is
done by BlueZ 5 D-Bus API.

BUG=chromium:329044
TEST=test_that --board ${BOARD} ${HOSTNAME} bluetooth_SDP_ServiceSearchRequestBasic

Change-Id: I2977617070ed6dd6fc0dbd2cfa08fd87e51c3c47
Reviewed-on: https://chromium-review.googlesource.com/187363
Tested-by: Artem Rakhov <arakhov@chromium.org>
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Commit-Queue: Artem Rakhov <arakhov@chromium.org>