power_status: Fix float division by zero exception
Some batteries have an issue with reporting the energy rate. Sometimes it is
zero which causes a ZeroDivisionError when we divide the total battery energy
by the energy rate. This fixes that by checking if the energy rate is non zero
before dividing. The only downside to this is that power_Idle will sometimes
report 0 for remaining time, but it will do this instead of fail the test.

BUG=chromium:346777
TEST=run power_LoadTest

Change-Id: Idce310bcd34db4e187d2e938fc143d17df964107
Reviewed-on: https://chromium-review.googlesource.com/187861
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>