bootperf_server: move client results to server results, don't copy
We keep raw results for each iteration of bootperf that we run.  We do
that by saving the results of each client test (BootPerf) to our own
per-iteration raw results folder.

With the old method of savings results (copy) things got confusing if
two sequential iterations of BootPerf didn't provide the same set of
files in the results directory.  If iteration .001 provided the files
("a", "b", "c") and iteration .002 provided ("b", "c", "d") then the
old method would include iteration.001's "a" in iteration.002's
results.

Fix this by just moving the client test results to the server test
results.  We'll avoid including the "keyval" file since the old code
tried not to include that in the raw results.

A nice side effect of this is that if the client includes a directory
in its results the move will handle it automatically.

BUG=chromium:259956
TEST=With a previous CL that added a directory, see that the directory
is logged OK.

Change-Id: I2c472dbddcf7ddad91370fc022873b31dedb7e27
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/187813
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>