Ensure that ui is stopped when monkeying with owner policy/key
The session_manager keeps copies of the owner key and policy in memory
and writes them to disk when it shuts down. This means that messing with them
while the UI is up will not persist across a UI restart. At least one test
was making this mistake (login_OwnershipTaken) leading to test-ordering
effects. This change updates ownership.py to correctly enforce this invariant.

BUG=chromium:338528
TEST=run login_*Ownership*

Change-Id: I72bb20b23abf3837e14b21cc6689cc383847b5db
Reviewed-on: https://chromium-review.googlesource.com/188077
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Mattias Nissler <mnissler@chromium.org>
Commit-Queue: Chris Masone <cmasone@chromium.org>