Touch FW test: do our own fingertip management for clicks
Previously the click gesture required that you start with no fingertips
selected and the gesture itself would take take of that for you.  This
was okay-ish because there wasn't a system in place to make sure the
robot had the right fingers.  Now, the touch FW test has progressed such
that we can tell it what fingers are expected for this gesture and it
will make sure that it's ready.  By letting the FW test take care of the
fingertip selection, we save on fingertip swaps between repeated
instances of the same gesture.  That is to say, between two runs of a
one-finger physical click, there's no reason to put the fingertip back,
and this makes that possible.

BUG=chromium:347737
TEST=manual testing

Change-Id: I9997058213300a05deb74c479cfa9c50cfa8fbd9
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188257
Reviewed-by: Dennis Kempin <denniskempin@chromium.org>