bootperf: Warn if we try to use shutdown data twice (and don't use it)
Copy the boot_id to the shutdown metrics when running
platform_BootPerf so we know if we reuse the same shutdown metrics
twice in a row.  Issue a warning if we detect this and then don't use
the data.

This is purposely not an error.  We should be detecting unclean or bad
shutdowns in some other test, not leveraging bootperf to do it.

BUG=chromium:259956
TEST=Run several times; echo bogus value into bootperf_ran and see
warning.

Change-Id: I3556f5010240d5d1589189ac4fabc3d1aeecaa02
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188177
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>