autotest (wifi): improve diagnostics for LowInitialBitrates
When we find a frame with an unexpectedly high bitrate, output
the timestamp of that frame (in local time). This will help
to find the same frame in wireshark.

To support this, change the tcpdump_analyzer code, to use
absolute times. This is because relative times are tricky
to get right, when combined with filters.

In particular, when given a filter expression, tcpdump -ttttt
will print frame times relative to the first matching frame.
This is not so useful when loading the same file in wireshark.

BUG=chromium:347690
TEST=manual (see below)

Manual test
1. ran test on peach_pi, where it currently fails
2. checked that the failure message included a timestamp for
   the frame, and that wireshark showed the referenced frame
   as being sent at a high bitrate
3. also ran network_WiFi_HiddenScan, network_WiFi_VisibleScan

Change-Id: I8c0b1a99c4a4b57293da597dbab176235c8fb563
Reviewed-on: https://chromium-review.googlesource.com/188215
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>