cros_host: Allow tests to specify a non-default reboot timeout
By default we do a 'reboot -f' after a reboot takes longer than 10
seconds.  For tests like bootPerf that are trying to _measure_ normal
reboot speeds, this is very non-ideal.  Let's fix it so that we can
override this default test-by-test.

BUG=chromium:259956
TEST=With future CL that makes bootperf request a timeout of 90,
added a printout of 'dargs['reboot_cmd']' in cros_host.py's
reboot() command. Saw:
  ((reboot & sleep 90; reboot -f &) </dev/null >/dev/null 2>&1 &)

Change-Id: I5c42088b2fa517980112fa9f53efb9c9d7e1de63
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188216
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>