bootperf_server: Give 90 seconds for normal reboot
We want to measure normal reboot, not a forced reboot.  Give the
reboot command a full 90 seconds (instead of 10) to run before we let
autotest try to do "reboot -f".

This is extra important because we measure reboot times based on
metrics that are written out by a normal reboot.

BUG=chromium:259956
TEST=Added a printout of 'dargs['reboot_cmd']' in cros_host.py's
reboot() command. Saw:
  ((reboot & sleep 90; reboot -f &) </dev/null >/dev/null 2>&1 &)

Change-Id: Ibac5b705ed31479ce60f36f837e8abb1ab96e176
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188217
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>