autotest: wait longer before querying shill on connection state after resume
Currently we wait for five seconds before we start to query shill on the
Service's state. This time would not suffice to determine the state of the
connection in few cases.

Devices using ath9k chipsets would not disassociate from AP before going to
suspend. After resume, the DUT would try to reuse connection from prior to
suspend. This works for a large population of APs. However, this has lead to a
confused state with some APs. In this state, the AP responds to actions frames
like Nullfunc or ADDBA request frames but does not respond to any data
frames like DHCP/ARP from the DUT. So lets wait for more time, 20 seconds,
before we start to query shill about the connection state.

BUG=chromium:346417
TEST=ran control.suspend in chaos lab on a DUT with atheros chipset against APs
(example ssid: dir_605l_g_ch5_wpa2). Check logs for link monitor error message
after resume, DUT reassociating to the AP and test passes with sucessful ping.

Change-Id: Id100cfcd73497c5f8b9b32078fc74e4e70acd90b
Signed-off-by: Ashok Nagarajan <asnagarajan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188361