[autotest] Expand provision test registration to other special tasks
We're looking to enable special actions to be taken during other special
tasks depending on the presense of certain labels.  This expands the
test registration out in a way that generalizes to more types of special
tasks better, and gives us a place to register the special actions for
verify, cleanup, and repair.  Note that reset is simply cleanup+verify,
so it gets no love here.

BUG=chromium:334418
TEST=doctest still passes

Change-Id: I17190b1bd095dd81d376432de8a78db5d95a87c8
Reviewed-on: https://chromium-review.googlesource.com/188453
Reviewed-by: Fang Deng <fdeng@chromium.org>
Tested-by: Alex Miller <milleral@chromium.org>
Reviewed-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Alex Miller <milleral@chromium.org>