Touch FW Test: fix 2f robot tap locations
The tap locations for 2f taps with the new robot were off-base.
Essentially they were a hold-over from the first robot which was
incapable of rotating.  As a result the 2f tap gestures were
approximated by simply tapping in different spots on the pad instead of
rotating the orientation of the two fingers.  The new robot is more than
capable of rotating, however, so this CL makes it actually tap at
different orientations and removes a little bit of cruft that's no
longer needed.

BUG=chromium:34049
TEST=I ran the test through all the tap/click gestures and they all
still worked and the 2f taps were at various angles instead of various
locations.

Change-Id: I07419b01d667ae225a35df05ed0d96407c2c6e56
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188327