Touch FW Test: shorten the settling time between gestures
It turns out that the extra 2 seconds that the test was giving the robot
to settle between gestures is over-kill.  The test itself takes some
time to process and the bug that this was originally attempting to fix
has since been fixed elsewhere.  This patch just shortens the setting
pause to half a second from 2 seconds, saving 1.5s per gesture, which
can actually add up to several minutes on a full test since there are a
lot of gestures, especially when you repeat them all 3 times.

BUG=none
TEST=I ran the test, and everything still function, there was just a
shorter pause and things moved a little faster

Change-Id: Ida8f9ee996aebd9891be1898babdc733abef2023
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188334
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>