Touch FW Test: Remove unneeded flag for clicks
The click.py script was modified to no longer need the
--manual-fingertips flag to suppress the gesture's fingertip
manipulations.  Now that is the default, so for the sake of cleanliness,
this CL removes the flag.  It works with the flag too, just because the
script ignores it, but it's better not to leave it there if it's not
getting used.

BUG=chromium:347737
TEST=manual testing
CQ-DEPEND=I4f26f3e1bef0f4e0a637269868392d56da7bdcbb

Change-Id: Ic02e5113edb8d502c13b02db42e74a78745abb92
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/188410
Reviewed-by: Dennis Kempin <denniskempin@chromium.org>