[Autotest] Modifications needed to make the tracker more flexible.
The phapi_client manages label queries and paging of results. This
allows us to crawl crbug. It is now also capable of skipping an expensive
issue creation operation by not retrieving the comments of the issue
if they're not needed.

The shadow config now only specifies the file name for tracker
credentials. The location of the file is composed based on the autotest
repo. This allows users who might not have an autotest repo mounted at
/usr/local/autotest to still use the reporter.

TEST=Ran a suite and filed/deduped bugs.
BUG=chromium:348040

Change-Id: I65ae6256e6b5c224eb5e8f17a96b6908eac338e0
Reviewed-on: https://chromium-review.googlesource.com/188527
Reviewed-by: Prashanth B <beeps@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>
Tested-by: Prashanth B <beeps@chromium.org>