power_DarkResume: add better failure detection and handling
This checks that the host actually turns back on at the end of the test. It
also turns the power supply back on before anything else in cleanup, so other
exceptions don't prevent that. We also check if the host is up before checking
boot IDs.

BUG=chrome-os-partner:26390
TEST=run power_DarkResumeShutdownServer in the lab

Change-Id: Id44b947022fa70c8d5cd04037b500a6bfeb2b473
Reviewed-on: https://chromium-review.googlesource.com/188942
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>