Remove site_linux_server
This class doesn't make sense in the new world with only a router
and a DUT.  We get most of the utility of what used to be the server
by creating delegate objects which understand how to use commands in
depth. For instance, netperf is run by the NetperfRunner rather than
being split between the client and the server.  NetperfRunner now
uses proxy_object.wifi_ip without worrying whether the proxy object
is a DUT or a router.  Netperf runner also looks up netperf commands
by itself rather than depending on the proxy object.

BUG=chromium:351555
TEST=All of wifi_matfunc passes with this change.

Change-Id: I093a488b0999cb6c418cab04ff02b3cf0f33136a
Reviewed-on: https://chromium-review.googlesource.com/189625
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>