Add new hardware_StorageWearoutDetect test
This test check wear out status for storage device available in SMART for SSD
and in ext_csd for eMMC version 5.0 or later. For previous version of eMMC,
it will be treat as data not available. The test will be failed if there is
SMART value is failing or eMMC wear out is in 90-100% band or higher.

CQ-DEPEND=CL:189600
BUG=chromium:351496
TEST=Tested in stout, pit, falco. Test result match expected result.

Change-Id: Idb183a05dd39160eea15efa4151ce3e44fbf69e0
Reviewed-on: https://chromium-review.googlesource.com/189626
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Puthikorn Voravootivat <puthik@chromium.org>