Fix race in MeasurementLogger
The MeasurementLogger runs in a separate thread. It runs in a loop that
periodically adds a timestamp to a list, then measurements to other lists. The
race condition is that we figure out a list of indices to use from the list of
timestamps. If we have added a timestamp to the list, but not the measurements,
there will be one too many indices for the list of measurements and a
ValueError will be thrown.

This fixes the problem by adding the measurement values then the timestamp.

BUG=chromium:318892
TEST=power_LoadTest

Change-Id: Icfd1c6b48e682824374711d94e0eb513e9057581
Reviewed-on: https://chromium-review.googlesource.com/190148
Reviewed-by: Todd Broch <tbroch@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>