firmware_TouchMTB: fix the gtk environment problems
After a tester performs a gesture, the mtplot screen is dumped
and should be displayed in the gesture capture window. However,
it failed to display the image file now due to some broken
configurations about gtk.

Specifically,
- Problem 1: The gtk library expects the GdkPixbuf image loader modules file
  at "/etc/gtk-2.0/gdk-pixbuf.loaders"
  which is actually at "/usr/local/etc/gtk-2.0/gdk-pixbuf.loaders"

- Problem 2: In the above GdkPixbuf image loader modules file, it
  specifies the *.so image modules to be at "/usr/lib*/..."
  which is actually at "/usr/local/lib*/..."

Problem 1 could be resolved by setting the environment variable
GDK_PIXBUF_MODULE_FILE to the proper path.

Problem 2 could be resolved by making a temporary symbolic link
when conducting the gesture recording process.

BUG=chromium:353874
TEST=Record some gestures in a chromebook machine.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ DISPLAY=:0 python main.py -m complete
Follow the prompts to record a few gestures, and press 'x' to quit.
Observe that the mtplot images of the performed gestures are
displayed in the X window each time a gesture is performed.

Change-Id: I1c38b374a3d9654582b8332c58e2694862dd94e1
Reviewed-on: https://chromium-review.googlesource.com/190996
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>