autotest_test: check its own crash_report status for logging_UserCrash
The logging_UserCrash test uses 'pgrep crash_reporter' to detect
whether the crash_reporter in its test finishes or not. It is
possible that the crash_reporter is invoked by some other test,
and in this case, the test will fail. After this change, it directly
checks the status of the process of crash_reporter invoked by this
test.

BUG=chromium:348855
TEST=run  logging_UserCrash when a crash_reporter process exists
     and it passes..
Change-Id: I8af61192e273e7dbd072f508107320c22142eb48
Reviewed-on: https://chromium-review.googlesource.com/190554
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Yunlian Jiang <yunlian@chromium.org>
Tested-by: Yunlian Jiang <yunlian@chromium.org>