Changing label for cellular repair devices
Created a new label (cellular_repair) to be used to run the
cellular_StaleModemReboot test. Added this label to the appropirate
devices and changing the label in the suite_scheduler to reflect this
change. also added a logging message to the test as well as another
modem state string.

BUG=chromium:349638
TEST=local test

Change-Id: I7541caf6bdb64d2c707284cb5e1d61cc36215bd0
Reviewed-on: https://chromium-review.googlesource.com/191118
Reviewed-by: Kris Rambish <krisr@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: harpreet Grewal <harpreet@chromium.org>
Tested-by: harpreet Grewal <harpreet@chromium.org>