Add some more logging for PDU fail
Saw an incident of PDU fail where the PDU that failed was
pingable and accessible though the sentry. We dont log the ping
packets for the rpm. We send only 2 packets and if there is any
loss raise an exception. May be one of them failed but we still
marked it as down. We can set a threshold for packet loss to
overcome this if that is the case. So logging the packets will
help debug this situation in the future.

BUG=None
TEST=ChaosPSK

Change-Id: If954ad83110561fb58eafa5a5fceb2d3123b326e
Reviewed-on: https://chromium-review.googlesource.com/191172
Tested-by: Bindu Mahadev <bmahadev@chromium.org>
Reviewed-by: Kris Rambish <krisr@chromium.org>
Commit-Queue: Kris Rambish <krisr@chromium.org>