autotest (networking): add test for disk full handling
Add a new server autotest which does the following:
a) starts a process to fill a filesystem on a DUT, and
   keep the file system full
b) runs a number of networking-focused client autotests
   on the device
c) checks the critical network processes keep running

While there: add a new client test which simply restarts
shill. This is inspired by the network_DefaultProfileServices
test. The reason for using a new test is that we cannot
expect network_DefaultProfileServices to pass when the
disk is full.

BUG=chromium:359772
TEST=manual (see below)

Manual test
-----------
Ran this new test with a ToT shill in a VM. Several failures
occur, because the networking daemons don't just handle a
full filesystem well:
- network_RestartShill.profile_missing:
  - triggers fail-safe
  - org.freedesktop.DBus.Error.ServiceUnknown exception
- many messages about metrics_daemon not running

Change-Id: I7d4d04dd38c72a61172c006627ae1f818e8e72d0
Reviewed-on: https://chromium-review.googlesource.com/191375
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>