Check for existing device specs before calibrating
Until now every time you ran a touch fw test with the robot, you would
have to manually recalibrate the device.  This invloved moving the
robot's fingers around to point to each of the corners of the touchpad
in turn so it can learn where the touch pad is, the size, etc.  The
calibration process take time and opens up a lot of uncertainly and the
possiblity of a miscalibrated device damaging the robot.

This CL adds in some code that checks in a couple places for existing
device specs before making the operator manually recalibrate.  The order
of preference is:
 * The local directory (in case you ran this before, it keeps the specs
   consistent across multiple runs)
 * The touchbot device_specs directory, where we will be uploading
   reference specs each time we test a new device so subsequent builds
   will already have one saved.
 * Manually recalibrate if all else fails

BUG=chromium:348174
TEST=I used this patch to kick off the test in all 3 scenarios and it
worked great.

Change-Id: Iddbe538c22ee7612fe053a250e7aad1f393899de
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/191525
Reviewed-by: Dennis Kempin <denniskempin@chromium.org>