Add power overrides to the unit test.
Powering on the APs can be very expensive in the unit test.
Since this is used for faster development the general workflow is:
  Manually power on AP
  Run unit test
  Make fixes
  Run unit test

Powering the AP off and on is not needed in this mode.  The user
can now adjust the OVERRIDE_POWER flag to turn this off and on.

TEST=This is a test
BUG=chromium:358299

Change-Id: I846a62439ef7ae023ab1dd67abb6fa5ea25857b4
Reviewed-on: https://chromium-review.googlesource.com/192391
Reviewed-by: Kris Rambish <krisr@chromium.org>
Commit-Queue: Tien Chang <tienchang@chromium.org>
Tested-by: Tien Chang <tienchang@chromium.org>