Resolving issue in western digital configurator.
Added additional checks to be sure that the page reloaded and all fields are
available for edit. Now we make sure that the radio for the band we choose is
enabled. I have also added a check to make sure that all the
elements in the login page are available before we hit submit.

BUG=chromium:354537,354531
TEST=None

Change-Id: Ib0e1ff6a9f52ae86c2ed300e92560d58431756e5
Reviewed-on: https://chromium-review.googlesource.com/192584
Reviewed-by: Kris Rambish <krisr@chromium.org>
Commit-Queue: Deepak Gopal <deepakg@chromium.org>
Tested-by: Deepak Gopal <deepakg@chromium.org>