Bluetooth SDP autotest: basic Service Attribute Request
Service Attribute Request is used to retrieve specified attribute values from
a specific service record. The service record handle of the desired service
record and a list of desired attribute IDs to be retrieved from that service
record are supplied as parameters.

This CL adds Service Attribute Request support to SDP Socket class,
as well as to all classes that are used to expose it's methods for autotest.
The CL also implements the very basic test for this type of request,
which is 'TP/SERVER/SA/BV-01-C' from SDP test specification.

BUG=chromium:329044
TEST=test_that --board ${BOARD} ${HOSTNAME} bluetooth_SDP_ServiceAttributeRequest

Change-Id: I4b236a331fea661584c396a2c3ff937bf91c2102
Reviewed-on: https://chromium-review.googlesource.com/196404
Tested-by: Artem Rakhov <arakhov@chromium.org>
Reviewed-by: Scott James Remnant <keybuk@chromium.org>
Commit-Queue: Artem Rakhov <arakhov@chromium.org>