firmware_TouchMTB: close gracefully when KeyboardInterrupt is captured
When the Keyboard Interrupt is captured, call the close() method
to exit FirmwareWindow object gracefully. This also allows the
firmware_TouchMTB object in main.py to do its own cleanup gracefully.

BUG=chromium:353874
TEST=Record some gestures in a chromebook machine.
$ cd /usr/local/autotest/tests/firmware_TouchMTB
$ DISPLAY=:0 python main.py -m complete
Follow the prompts to record a few gestures, and press ctrl-c
to interrupt the tests. Observe that it shows the message:

Display the html test report on the browser.
This may take a while...

... (followed by more messages)

Change-Id: I79fbf3d563d4a9fd24aa36c2e36581597d02db96
Reviewed-on: https://chromium-review.googlesource.com/192713
Reviewed-by: Charles Mooney <charliemooney@chromium.org>
Tested-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: Shyh-In Hwang <josephsih@chromium.org>