[autotest] For suites, merely warn when a label isn't found.
Previously, if a request was made (e.g. via run_suite) to run a test
for a suite, but test dependencies specified a non-existent label,
the entire suite would fail.  This changes the behavior to catch the
exception and report a warning instead; the behavior is similar to
what we do when a test can't be scheduled for valid labels for which
no host meets the criteria.

BUG=chromium:353541
TEST=run_suite on a local instance

Change-Id: I969c00b630ba287718e93eaf9e76004417f961c6
Reviewed-on: https://chromium-review.googlesource.com/192867
Tested-by: Richard Barnette <jrbarnette@chromium.org>
Reviewed-by: Dan Shi <dshi@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Richard Barnette <jrbarnette@chromium.org>