[autotest] Include parent_job_id as a tie breaker in the host request.
Teach the rdb to give hosts to requests from the same suite by including
parent_job_id in the acquire hosts request, and using it to order
requests of equal priority. This means that we will only group requests
from the same suite, which will result in an O(number of concurrent suites)
increase in database queries to find hosts, but it should lead to an
overall win since suites will timeout less often if they're allowed to
run till completion before hosts in the same pool are re-allocated to another
suite.

TEST=Ran suites, added unittests that failed before and passes now.
BUG=chromium:351861

Change-Id: Ia6e01f92484fb7bfd2e2e1e780cc4a16832fe901
Reviewed-on: https://chromium-review.googlesource.com/192965
Tested-by: Prashanth B <beeps@chromium.org>
Reviewed-by: Alex Miller <milleral@chromium.org>
Commit-Queue: Prashanth B <beeps@chromium.org>