Revert "Add tests to Smoke/BVT suite."
VMTests are failing on canaries due to the stateful partition getting filled.
I'm not convinced these tests are directly at fault; it may simply be that
there's a cumulative effect. THe patch reverted by this CL seemed to kick us 
over the edge, though, as the first canary that failed in this way kicked off 
about 10 minutes after it CL landed.

BUG=359422
TEST=None

This reverts commit 681fefb68967d6bfbbc31004cbd61c8c21c57062.

Change-Id: I02216e4f5ce7b14ecbb15c5ee65188a66fe3c575
Reviewed-on: https://chromium-review.googlesource.com/192978
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>